refactor: 코드 분리 (FE-55)
