+++ src/components/task-list-page/Task.tsx
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import { DateTask, PatchTaskRequest } from "@coworkers-types";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import classNames from "classnames";
